requirements.txt
streamlitpandasnumpysklearn-contrib-python-headless